LayoutTests/imported/w3c/web-platform-tests/css/css-ui/appearance-initial-value-001-expected.txt


FAIL support for appearance assert_equals: expected "button" but got ""
FAIL initial value for appearance assert_equals: expected "none" but got ""
PASS support for -webkit-appearance 
PASS initial value for -webkit-appearance 
